backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
greenlet
pyyaml
requests
websockets
sqlalchemy
aiosqlite
pydantic
python-multipart
